Ухх. Вот попарился (в очередной раз) с коллизиями. Теперь вроде все обрабатывается адекватно. Можно будет весь день теперь заниматься собственно геймплеем. Основной каркас для этого уже точно готов. Думаю будет получаться что-то вроде канабалта. Надо еще подумать можно ли прикрутить управление только одной клавишей.. ведь в иделе хотелось бы еще сделать подкад кроме прыжка.. Кстати вот с чего надо будет начать: сделать прыжок

+++ game/CGameBg.h
@@ -9,7 +9,7 @@ class CGameBg: public CDisplayObject
     public:
         CGameBg();
         virtual ~CGameBg();
-        void parallax(int delta);
+        void parallax(float delta);
     protected:
     private:
         float d_layer0;
